Create custom_nms.py
this file includes a custom nms layer - which uses inter over area instead of inter over union as the anchor selection metric. The issue of crowns of different types overlapping each other is also solved. This change inceases the speed and mask quality a bit.
Another function called nms_mask is used in prediction part and is part of the proprocessing where instead of eliminating all crowns with a low threshold, the nms is used and the nms metric is no longer box iou but mask iou. This change increases the output quality a lot.